🔒 Added protection against CSRF by the use of CSRF tokens

+++ src/controllers/account/destroy.php
@@ -1,16 +1,23 @@
 <?php
 
+if(isset($_POST["token"]) && validToken($_POST["token"])){
     deleteAccount(Session::get("email"));
+    Session::destroy();    
+}
 
-    Session::destroy();
+redirect("/");
 
-    redirect("/");
+function validToken($token)
+{
+    return $token === $_SESSION["token"];
+}
 
-    function deleteAccount($email){
+function deleteAccount($email)
+{
 
-        $dbConfig = require_once base_path("src/db/configurations/dbconfig.php");
-        
-        $db = new DeleteDbHandler($dbConfig, "db-username", "db-password");
-        $db -> delete($email);
+    $dbConfig = require_once base_path("src/db/configurations/dbconfig.php");
 
-    }
+    $db = new DeleteDbHandler($dbConfig, "db-username", "db-password");
+    $db->delete($email);
+
+}
